refactor: contentwindow stuff moved to styles folder
